requirements.txt
aiogram==3.*
aiosqlite==0.20.*
python-dotenv==1.*
